move the generator functions for imgCIF entries to separate file
